move utils to own dir
